requirements.txt
streamlit
pandas
scikit-learn
numpy
matplotlib
seaborn
plotly
Pillow
